feat:Setting up navigation
Configuration for navigation into a new App.js
With multiple layout in the screens folder and a
global file of styles to apply to them
