+++ screens/About.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import { Text, View } from 'react-native'
+import { global } from '../styles/Global'
+
+const About = () => {
+  return (
+    <View style={global.container}>
+      <Text>About</Text>
+    </View>
+  )
+}
+
+export default About
