+++ screens/Home.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import { Text, View } from 'react-native'
+import { global } from '../styles/Global'
+
+const Home = () => {
+  return (
+    <View style={global.container}>
+      <Text>Home</Text>
+    </View>
+  )
+}
+
+export default Home
