+++ screens/Tasks.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import { Text, View } from 'react-native'
+import { global } from '../styles/Global'
+
+const Tasks = () => {
+  return (
+    <View style={global.container}>
+      <Text>Tasks</Text>
+    </View>
+  )
+}
+
+export default Tasks
