+++ styles/Global.js
@@ -1,0 +1,7 @@
+import { StyleSheet } from "react-native";
+
+export const global = StyleSheet.create({
+  container:{
+    padding: 30
+  }
+})
